% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection_function.R
\name{bakerbat_database_connect}
\alias{bakerbat_database_connect}
\title{This function simplifies connecting with the bakerbat_database. It calls the username
and password saved as a .txt file with username as the first line and password as the second line.
Make sure there is not superfluous white space in this text file.}
\usage{
bakerbat_database_connect(username.password.file.path, suppress.bat = T)
}
\arguments{
\item{username.password.file.path}{The path of the text file containing the username and password to the database account}

\item{suppress.bat}{Set to TRUE}
}
\value{
The database connection object
}
\description{
This function simplifies connecting with the bakerbat_database. It calls the username
and password saved as a .txt file with username as the first line and password as the second line.
Make sure there is not superfluous white space in this text file.
}
