% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection_function.R
\name{bakerbat_database_disconnect}
\alias{bakerbat_database_disconnect}
\title{Disconnect from the bakerbat_database}
\usage{
bakerbat_database_disconnect(database.connection.object)
}
\arguments{
\item{database.connection.object}{The database connection object created with the bakerbat_database_connect() function}
}
\value{
message indicating disconnecting from bakerbat_database
}
\description{
Disconnect from the bakerbat_database
}
