% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_query_function.R
\name{Bat_Data}
\alias{Bat_Data}
\title{Bat Data function
This function queries general data of bats from the Bakerbat project}
\usage{
Bat_Data(database.connection.object)
}
\arguments{
\item{database.connection.object}{The database connection object created with the new_preempt_database_connect() function}
}
\value{
a dataframe
}
\description{
Bat Data function
This function queries general data of bats from the Bakerbat project
}
