% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_query_function.R
\name{census_bat_data}
\alias{census_bat_data}
\title{census bat data function
This function queries general data of bats from the Bakerbat project}
\usage{
census_bat_data(database.connection.object)
}
\arguments{
\item{database.connection.object}{The database connection object created with the new_preempt_database_connect() function}
}
\value{
a dataframe
}
\description{
census bat data function
This function queries general data of bats from the Bakerbat project
}
