% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_query_function.R
\name{census_sample_extracted}
\alias{census_sample_extracted}
\title{sample extracted function
This function queries details on the type of samples extracted during the Bakerbat project}
\usage{
census_sample_extracted(database.connection.object)
}
\arguments{
\item{database.connection.object}{The database connection object created with the new_preempt_database_connect() function}
}
\value{
a dataframe
}
\description{
sample extracted function
This function queries details on the type of samples extracted during the Bakerbat project
}
