% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_query_function.R
\name{census_tattoo_information}
\alias{census_tattoo_information}
\title{tattoo information function
This function queries details of the tattooing activities during the Bakerbat project}
\usage{
census_tattoo_information(database.connection.object)
}
\arguments{
\item{database.connection.object}{The database connection object created with the new_preempt_database_connect() function}
}
\value{
a dataframe
}
\description{
tattoo information function
This function queries details of the tattooing activities during the Bakerbat project
}
