% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_query_function.R
\name{list_bakerbat_database_tables}
\alias{list_bakerbat_database_tables}
\title{List Tables Available}
\usage{
list_bakerbat_database_tables(database.connection.object)
}
\arguments{
\item{database.connection.object}{The database connection object created with the preempt_database_connect() function}
}
\value{
A vector of all tables in the bakerbat_database.
}
\description{
This function lists all currently available tables in the bakerbat_database.
}
\keyword{database}
\keyword{tables}
