% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_query_function.R
\name{maternal_colony_exp_meta_data}
\alias{maternal_colony_exp_meta_data}
\title{maternal colony experiment - metadata function
This function queries details on the metadata of the maternal colony experiment}
\usage{
maternal_colony_exp_meta_data(database.connection.object)
}
\arguments{
\item{database.connection.object}{The database connection object created with the new_preempt_database_connect() function}
}
\value{
a dataframe
}
\description{
maternal colony experiment - metadata function
This function queries details on the metadata of the maternal colony experiment
}
