% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_query_function.R
\name{maternal_colony_exp_pup_registry}
\alias{maternal_colony_exp_pup_registry}
\title{maternal colony experiment - registry function
This function queries details of the birthdays of pups from the maternal colony experiment}
\usage{
maternal_colony_exp_pup_registry(database.connection.object)
}
\arguments{
\item{database.connection.object}{The database connection object created with the new_preempt_database_connect() function}
}
\value{
a dataframe
}
\description{
maternal colony experiment - registry function
This function queries details of the birthdays of pups from the maternal colony experiment
}
