% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_query_function.R
\name{maternal_colony_exp_pup_sample_extracted}
\alias{maternal_colony_exp_pup_sample_extracted}
\title{maternal colony experiment - pup sample extracted function
This function queries details on the type of samples extracted from the pups of the maternal colony experiment}
\usage{
maternal_colony_exp_pup_sample_extracted(database.connection.object)
}
\arguments{
\item{database.connection.object}{The database connection object created with the new_preempt_database_connect() function}
}
\value{
a dataframe
}
\description{
maternal colony experiment - pup sample extracted function
This function queries details on the type of samples extracted from the pups of the maternal colony experiment
}
