% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_query_function.R
\name{maternal_colony_exp_technitians}
\alias{maternal_colony_exp_technitians}
\title{maternal colony experiment - technitians function
This function queries details of the activities of each technitian during the maternal colony experiment}
\usage{
maternal_colony_exp_technitians(database.connection.object)
}
\arguments{
\item{database.connection.object}{The database connection object created with the new_preempt_database_connect() function}
}
\value{
a dataframe
}
\description{
maternal colony experiment - technitians function
This function queries details of the activities of each technitian during the maternal colony experiment
}
