% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_query_function.R
\name{mom_pup_pair}
\alias{mom_pup_pair}
\title{maternal colony experiment - mom-pup-pair function
This function shows details of the mom-pup pairs from the maternal colony experiment}
\usage{
mom_pup_pair(database.connection.object)
}
\arguments{
\item{database.connection.object}{The database connection object created with the new_preempt_database_connect() function}
}
\value{
a dataframe
}
\description{
maternal colony experiment - mom-pup-pair function
This function shows details of the mom-pup pairs from the maternal colony experiment
}
