% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_query_function.R
\name{RIP}
\alias{RIP}
\title{RIP function
This function queries the id and date when the bats passed}
\usage{
RIP(database.connection.object)
}
\arguments{
\item{database.connection.object}{The database connection object created with the new_preempt_database_connect() function}
}
\value{
a dataframe
}
\description{
RIP function
This function queries the id and date when the bats passed
}
