% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_query_function.R
\name{Sample_Extracted}
\alias{Sample_Extracted}
\title{Sample Extracted function
This function queries details on the type of samples extracted during the Bakerbat project}
\usage{
Sample_Extracted(database.connection.object)
}
\arguments{
\item{database.connection.object}{The database connection object created with the new_preempt_database_connect() function}
}
\value{
a dataframe
}
\description{
Sample Extracted function
This function queries details on the type of samples extracted during the Bakerbat project
}
