% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_query_function.R
\name{Tattoo_Information}
\alias{Tattoo_Information}
\title{Tattoo Information function
This function queries details of the tattooing activities during the Bakerbat project}
\usage{
Tattoo_Information(database.connection.object)
}
\arguments{
\item{database.connection.object}{The database connection object created with the new_preempt_database_connect() function}
}
\value{
a dataframe
}
\description{
Tattoo Information function
This function queries details of the tattooing activities during the Bakerbat project
}
