% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_query_function.R
\name{technitians}
\alias{technitians}
\title{technitians}
\usage{
technitians(database.connection.object)
}
\arguments{
\item{database.connection.object}{The database connection object created with the bakerbat_database_connect() function}
}
\value{
A vector of all tables in the bakerbat_database.
}
\description{
This function queries census_technitians + technitians + activity_roles tables to create the final Technitians table
}
\keyword{database}
\keyword{tables}
