% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_query_function.R
\name{Census_Meta_Data}
\alias{Census_Meta_Data}
\title{Census_Meta_Data}
\usage{
Census_Meta_Data(database.connection.object)
}
\arguments{
\item{database.connection.object}{The database connection object created with the bakerbat_database_connect() function}
}
\value{
A vector of all tables in the bakerbat_database.
}
\description{
This function queries census_meta_data + census_technitians + technitians tables to create the final Census_Meta_Data table
}
\keyword{database}
\keyword{tables}
